requirements.txt
asttokens==2.4.1
decorator==5.1.1
executing==2.0.1
iniconfig==2.0.0
ipython==8.19.0
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
parsy==2.1
pexpect==4.9.0
pluggy==1.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==7.4.4
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
wcwidth==0.2.12
pytest
python-dateutil
python-fasthtml
